+++ Client/src/Components/DatasetUpload/DatasetForm/DataSection/DatasetDataTable.tsx
@@ -2,15 +2,15 @@ import 'react-data-grid/dist/react-data-grid.css'
 
 import { Box, Button, Grid, Typography } from '@material-ui/core'
 import DataGrid, { SelectColumn, TextEditor } from 'react-data-grid'
-import { IContent, IData, IVariable, newVariable } from '../../../Models/Datasets/IDatasetModel'
+import { IContent, IData, IVariable, newVariable } from '../../../../Models/Datasets/IDatasetModel'
 import React, { useState } from 'react'
 
 import { EditVaraibleModal } from './EditVariableModal'
 import { VariableHeader } from './VariableHeader'
-import { decorateDataErrors } from '../../../Common/Helpers/DatasetErrorDecorator'
-import { useDimensions } from '../../Utils/Hooks/useDimensions'
+import { decorateDataErrors } from '../../../../Common/Helpers/DatasetErrorDecorator'
+import { useDimensions } from '../../../Utils/Hooks/useDimensions'
 import { useFormikContext } from 'formik'
-import { useVariableNames } from '../../Utils/Hooks/useVariableNames'
+import { useVariableNames } from '../../../Utils/Hooks/useVariableNames'
 
 interface IProps {
   data: IData,
