+++ Client/src/Components/DatasetUpload/DatasetForm/ReferenceSection/AuthorsList.tsx
@@ -1,5 +1,5 @@
 import { Grid, IconButton } from "@material-ui/core"
-import { IAuthor, newAuthor } from "../../../Models/Datasets/IDatasetModel"
+import { IAuthor, newAuthor } from "../../../../Models/Datasets/IDatasetModel"
 
 import AddIcon from '@material-ui/icons/Add'
 import { ArrayHelpers } from "formik"
